# gawk-admin/openapi.yaml
...
openapi: 3.1.1
info:
  title: gawk-admin API
  version: 1.3.0 # x-release-please-version
  summary: Moderation and fleet-operations API for a gawk deployment.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    The HTTP API of `gawk-admin`, the moderation portal of a
    [gawk](https://github.com/Tuhis/gawk) deployment. The portal SPA uses it,
    and external integrations such as bots use the same API.

    ## Authentication

    Every operation except `GET /api/v1/openapi.json` requires a bearer JWT
    issued by the deployment's OIDC provider:

    ```
    Authorization: Bearer <token>
    ```

    Each operation lists the role it requires under `x-gawk-roles`. A valid
    token without that role gets `403`. There are no cookies and no
    server-side session.

    The served copy of this document names the role claim value this
    deployment is configured with (`-operator-role`). The repository copy uses
    the symbolic name `operator`.

    Service identities use the client-credentials grant against the same
    provider. `docs/self-hosting.md` §9.8 has a worked example.

    The routes that need no token: `GET /auth/config` returns the issuer,
    client ID and audience a browser client needs; `GET /api/v1/openapi.json`
    returns this document; `GET /api/v1/asyncapi.json` and
    `GET /api/v1/schemas/events/{name}` return the event contract.

    ## Compatibility

...
    treat unknown enum values as unknown rather than as errors.

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/bans:
    get:
      operationId: listBans
      tags: [bans]
      summary: List bans
      description: |
        `state=active` (the default) returns the whole active set in one
        response. `state=all` returns the full history, newest first, paged by
        the `(afterCreatedAt, afterId)` cursor. Send both cursor parameters
        or neither.

        `nextAfter` is always present. It is `null` on the last page, and
        always `null` for `state=active`.
      x-gawk-roles: [operator]
      parameters:
        - name: state
          in: query
          required: false
          description: Which set to return.
          schema:
            type: string
            enum: [active, all]
            default: active
        - name: limit
          in: query
          required: false
          description: Page size for `state=all`. Values above 500 are clamped; zero or negative means the default.
          schema: { type: integer, format: int32, default: 50, maximum: 500 }
        - name: afterCreatedAt
          in: query
          required: false
          description: The previous page's `nextAfter.createdAt`. Must be sent together with `afterId`.
          schema: { type: string, format: date-time }
        - name: afterId
          in: query
          required: false
          description: The previous page's `nextAfter.id`. Must be sent together with `afterCreatedAt`.
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: One page of bans.
          x-gawk-sensitive: true
          x-gawk-sensitive-reason: A ban target may be a raw broadcast ID or a publisher CIDR.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BansPage' }
              example:
                bans:
                  - id: 1f1c2b3a-4d5e-4f60-8a91-0b2c3d4e5f60
                    target: { type: ip, value: 203.0.113.0/24 }
                    state: active
                    reason: evading an ID ban
                    createdAt: '2026-09-16T09:31:00Z'
                    createdBy: operator@example.org
                    expiresAt: null
                    crName: ip-203-0-113-0-24
                nextAfter: null
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/Unavailable' }
    post:
      operationId: createBan
      tags: [bans]
      summary: Create a ban
      description: |
        Bans a broadcast ID or a publisher address range.

        For an `ip` ban, `target.value` may be an address, a CIDR, or the
        literal `publisher`. The `publisher` form resolves the live
        publisher's address server-side from `sourceBroadcastId`, which is
        then required.

        A bare address is widened to `prefixLength`, which defaults to `/32`
        for IPv4 and `/64` for IPv6. A `prefixLength` that disagrees with a
        CIDR in `value` is rejected.

        `expiresAt: null` (or absent) makes the ban permanent.

        See "Ban mutations" above for the meaning of `201` versus `202`.
      x-gawk-roles: [operator]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBanRequest' }
            example:
              target: { type: ip, value: publisher, prefixLength: 24 }
              expiresAt: '2026-10-16T09:31:00Z'
              reason: evading an ID ban
              sourceBroadcastId: ABC234
      responses:
        '201':
          description: The ban is recorded and enforced.
          x-gawk-sensitive: true
          x-gawk-sensitive-reason: The ban target may be a raw broadcast ID or a publisher CIDR.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ban' }
              example:
                id: 2a2d3e4f-5061-4728-9b3c-4d5e6f708192
                target: { type: ip, value: 203.0.113.0/24 }
                state: active
                reason: evading an ID ban
                createdAt: '2026-09-16T09:31:00Z'
                createdBy: operator@example.org
                expiresAt: '2026-10-16T09:31:00Z'
                sourceBroadcastId: ABC234
                crName: ip-203-0-113-0-24
        '202':
          description: The ban is recorded; enforcement is pending. Do not retry.
          x-gawk-sensitive: true
          x-gawk-sensitive-reason: The ban target may be a raw broadcast ID or a publisher CIDR.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ban' }
              example:
                id: 2a2d3e4f-5061-4728-9b3c-4d5e6f708192
                target: { type: ip, value: 203.0.113.0/24 }
                state: active
                reason: evading an ID ban
                createdAt: '2026-09-16T09:31:00Z'
                createdBy: operator@example.org
                expiresAt: null
                enforcement:
                  inSync: false
                  detail: >-
                    The ban is recorded but NOT enforced yet — its Kubernetes
                    enforcement object could not be written; the reconciler
                    retries within a minute, so do not re-submit.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '409':
          description: An active ban already covers this target.
          x-gawk-sensitive: true
          x-gawk-sensitive-reason: The ban target may be a raw broadcast ID or a publisher CIDR.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BanConflict' }
              example:
                error:
                  code: duplicate_active
                  message: an active ban already covers this target
                ban:
                  id: 2a2d3e4f-5061-4728-9b3c-4d5e6f708192
                  target: { type: ip, value: 203.0.113.0/24 }
                  state: active
                  reason: evading an ID ban
                  createdAt: '2026-09-16T09:31:00Z'
                  createdBy: operator@example.org
                  expiresAt: null
        '500': { $ref: '#/components/responses/Internal' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An access token issued by the deployment's OIDC provider. Its role
        claim must include the role the operation lists under `x-gawk-roles`.
  responses:
    BadRequest:
      description: The request is malformed or a target could not be resolved.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: bad_request
              message: reason is required
    Unauthorized:
      description: The bearer token is missing or invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: unauthorized
              message: a valid bearer token is required
    Forbidden:
      description: The token lacks the required role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: forbidden
              message: this account does not hold the operator role
    TooManyRequests:
      description: Too many rejected tokens from this address.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer, format: int32 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: rate_limited
              message: too many invalid credentials; slow down
...
    Internal:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: internal
              message: internal error
    Unavailable:
      description: A dependency is unreachable; retry later.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error:
              code: unavailable
              message: the moderation database is unreachable; enforcement of existing bans is unaffected
  schemas:
    ErrorCode:
      type: string
      description: |
        Machine-readable error code. The enum only grows; treat unknown values
        as unknown. The first four are issued by the authentication layer and
        can occur on any operation that requires a role.
      enum:
        # From the authentication layer, on any operation with a role.
        - unauthorized
        - idp_unavailable
        - forbidden
        - rate_limited
        # From the handlers.
        - bad_request
        - not_found
        - duplicate_active
        - duplicate_name
        - source_immutable
        - internal
        - unavailable
        - invalid_target
        - ban_not_active
        - room_exists
        - room_not_static
        - room_not_dynamic
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          description: Human-readable explanation, for display only.
    Error:
      type: object
      description: The error envelope every failing operation returns.
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
...
    BroadcastId:
      type: string
      description: |
        A raw, joinable broadcast ID: six characters of the broadcast
        alphabet. Case-insensitive on input, normalized on output.
      example: ABC234
...
    BanTarget:
      type: object
      description: What a ban applies to.
      required: [type, value]
      properties:
        type:
          type: string
          enum: [broadcastId, ip]
        value:
          type: string
          description: A broadcast ID, or a CIDR for an `ip` ban. Always a prefix on output.
    Enforcement:
      type: object
      description: |
        Whether the Kubernetes enforcement object matches the record. Present
        only on a `202`; absent means the two are in sync.

        A pending create means the ban is recorded but not yet enforced. A
        pending removal means the ban is lifted in the record but the target
        is still enforced.
      required: [inSync]
      properties:
        inSync:
          type: boolean
        detail:
          type: string
          description: One sentence describing the pending direction. Suitable for display.
    Ban:
      type: object
      description: One ban record.
      required: [id, target, state, reason, createdAt, createdBy, expiresAt]
      properties:
        id: { type: string, format: uuid }
        target: { $ref: '#/components/schemas/BanTarget' }
        state:
          type: string
          enum: [active, expired, removed]
        reason:
          type: string
          description: Operator-private context. Never included in outbound events.
        createdAt: { type: string, format: date-time }
        createdBy:
          type: string
          description: The identity that created the ban, or `system`.
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: Always present. `null` means the ban is permanent.
        removedAt: { type: string, format: date-time }
        removedBy: { type: string }
        sourceBroadcastId:
          allOf:
            - $ref: '#/components/schemas/BroadcastId'
          description: The broadcast the ban was taken against, if any.
        crName:
          type: string
          description: Name of the Kubernetes enforcement object. Derived from the target.
        enforcement: { $ref: '#/components/schemas/Enforcement' }
    BanConflict:
      type: object
      description: A `409` that returns the active ban already covering the target.
      required: [error, ban]
      properties:
        error: { $ref: '#/components/schemas/ErrorBody' }
        ban: { $ref: '#/components/schemas/Ban' }
    BanCursor:
      type: object
      description: |
        Cursor for ban history. Both halves are needed: `createdAt` alone is
        not unique and `id` alone does not order.
      required: [createdAt, id]
      properties:
        createdAt:
          type: string
          format: date-time
          description: RFC 3339 with nanoseconds. Send it back unchanged.
        id: { type: string, format: uuid }
    BansPage:
      type: object
      required: [bans, nextAfter]
      properties:
        bans:
          type: array
          items: { $ref: '#/components/schemas/Ban' }
        nextAfter:
          oneOf:
            - $ref: '#/components/schemas/BanCursor'
            - type: 'null'
          description: Cursor for the next page, or `null` on the last page.
    CreateBanRequest:
      type: object
      required: [target, reason]
      properties:
        target:
          type: object
          required: [type, value]
          properties:
            type:
              type: string
              enum: [broadcastId, ip]
            value:
              type: string
              description: |
                A broadcast ID, an IP address, a CIDR, or the literal
                `publisher` for an `ip` ban. `publisher` resolves the live
                publisher's address from `sourceBroadcastId`, which is then
                required.
            prefixLength:
              type: integer
              format: int32
              description: |
                Prefix length for an IP ban. Defaults to `32` for IPv4 and
                `64` for IPv6. Rejected if it disagrees with a CIDR in `value`.
        expiresAt:
          type: [string, 'null']
          format: date-time
          description: RFC 3339. `null` or absent means permanent.
        reason:
          type: string
          maxLength: 512
          description: Operator-private context. Never included in outbound events.
        sourceBroadcastId:
          allOf:
            - $ref: '#/components/schemas/BroadcastId'
          description: The broadcast this ban stems from. Required when `target.value` is `publisher`.
...